Move other helpers into module
